now can query urlpath from transform catalogs
urlpath is not an arg for the transform catalogs, but users want to be able to see what they are for the target source. Now the transform is followed through to the target with `cat.follow_target()` which also sets the urlpath attribute in the target source from the orig source. This is run in mc.setup() and mc.find_availability so that it is always updated. mc.select_date_range() runs `cat.update_urlpath` to bring the newly discovered urlpath list to the target and makes it discoverable like in the other cases too with `source.urlpath`.