warning for date change
previously, if a kernel was still active for a user from the day before, and then was used, user parameters would not update and a to_dask command might not work because the yesterday or today parameter was set the day before and would be out of date. The behavior is the same now, but there are checks when to_dask is called to check to see if this is the case and warn the user if it is.