new catch for if time slice not available
Currently the NCEI archive aggregation appears to be missing some time steps. In any case, now if a time slice is requested that is not available from a static link (otherwise it would come up earlier in the process), a warning will be posted and the dataset without time slice will be returned.