changed CAT_PATH and forcing single threading
CAT_PATH in __init__ now points to the support_files subdirectory in the model_catalogs package, which is where the transform.yaml file can be found. Otherwise I do not think it is used anymore.
Also there is an issue when open_mfdataset is used (which is for our noagg sources) in which you get an error the first time you try to read in a file but not if you try it again. You can pin netcdf or you can do this which is enforce single threading